requirements.txt
absl-py==1.2.0
apturl==0.5.2
attrs==22.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cvzone==1.5.6
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
fonttools==4.37.1
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gTTS==2.2.4
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.3
mediapipe==0.8.11
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
playsound==1.3.0
protobuf==3.12.4
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pygame==2.1.2
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySimpleGUI==4.60.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-xlib==0.15
pyttsx==1.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
scipy==1.9.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sounddevice==0.4.5
systemd-python==234
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wavio==0.0.4
xdg==5
xkit==0.0.0
zipp==1.0.0
